requirements.txt
google-cloud-aiplatform[reasoningengine]>=1.49.0
google-cloud-storage>=2.16.0
google-adk==1.26.0
google-genai>=0.7.1
python-dotenv>=1.0.1
